# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Booking Platformasi ucun Backend API. Randevu vaxtlari is qrafiki, nahar fasilesi ve slot addimi esasinda runtime-da hesablanir; bildirisler WebSocket ve FCM ile realtime catdirilir.",
        "title": "Booking Service API",
        "contact": {},
        "version": "2.0"
    },
    "host": "booking-service-sld9.onrender.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/staff": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all staff members for the authenticated business. Returns complete list of active and inactive staff with user associations, specializations, and availability status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "List All Staff Members",
                "responses": {
                    "200": {
                        "description": "Staff list retrieved successfully (array of StaffWithUserHTTPResponse)",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated or business_id missing",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new staff member profile for multi-staff business. Used when directly adding staff without invitation flow (admin creates profile). Staff member becomes available for service assignment. Only available for multi-staff businesses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "Create Staff Profile",
                "parameters": [
                    {
                        "description": "Staff profile data (FirstName, LastName, Email, Phone, Specializations - optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.CreateStaffHTTPRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Staff profile created successfully with generated UUID",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error - invalid or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated or business_id missing",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/staff/invites": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends invitation to join business as staff member (zero-knowledge invitation flow). Generates unique invitation token valid for 7 days. Invitee receives token and can accept by creating password and joining business. Used for multi-staff businesses to onboard team members.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "Invite Staff Member",
                "parameters": [
                    {
                        "description": "Invitation details (FirstName, LastName, Email, Phone, Role - provider_owner, staff, customer)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.InviteStaffHTTPRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invitation created successfully with token and expiration",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error - invalid email format or staff already invited",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated or business_id missing",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/staff/invites/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Completes staff member onboarding by accepting invitation and setting password. User must be authenticated (user_id from context). Uses zero-knowledge invitation flow - staff sets their own password without server storing temporary passwords. Staff becomes active member after acceptance. Token is marked as used and cannot be reused.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "Accept Invitation (Staff Onboarding)",
                "parameters": [
                    {
                        "description": "Invitation token and password for account setup",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.AcceptInviteHTTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation accepted successfully, staff member activated",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or expired token, password validation failure",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated or user_id missing",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/staff/invites/validate": {
            "post": {
                "description": "Validates staff invitation token before acceptance. Returns invitation details if token is valid and not expired. No authentication required - used during invitation onboarding flow. Token must be valid and within 7-day expiration window.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Staff"
                ],
                "summary": "Validate Invitation Token",
                "parameters": [
                    {
                        "description": "Invitation token to validate",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ValidateInviteHTTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token is valid with invitation details (FirstName, LastName, Email, BusinessName)",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid, expired, or already-used invitation token",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers_staff.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_http_handlers_staff.AcceptInviteHTTPRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers_staff.CreateStaffHTTPRequest": {
            "type": "object",
            "properties": {
                "location_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers_staff.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_http_handlers_staff.InviteStaffHTTPRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "location_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers_staff.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_http_handlers_staff.ValidateInviteHTTPRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}